+++ WithoutHaste.Sequences/WithoutHaste.Sequences/SexyPrime.cs
@@ -11,12 +11,12 @@ namespace WithoutHaste.Sequences
 	public class SexyPrime : RelativePrime
 	{
 		/// <inheritdoc/>
-		internal static new int[] TestNumbers = new int[] { 5, 7, 11, 13, 17, 19, 23, 29, 31, 37, 41, 43, 47, 53, 59, 61, 67, 73, 79, 83, 89, 97, 101, 103, 107, 109, 113, 131, 137, 151, 157, 163, 167, 173, 179, 191, 193, 197, 199 };
+		internal static new long[] TestNumbers = new long[] { 5, 7, 11, 13, 17, 19, 23, 29, 31, 37, 41, 43, 47, 53, 59, 61, 67, 73, 79, 83, 89, 97, 101, 103, 107, 109, 113, 131, 137, 151, 157, 163, 167, 173, 179, 191, 193, 197, 199 };
 
 		/// <inheritdoc/>
 		protected override int X { get { return 6; } }
 
-		public SexyPrime(int max) : base(max)
+		public SexyPrime(long max) : base(max)
 		{
 		}
 	}
